Add ReportGenerator and GitVersion.Tool to build project
In this commit, two new package references have been added to the _build.csproj project file. The first added package is "ReportGenerator" with version 5.2.1, which is useful for generating reports on code coverage. The second addition is "GitVersion.Tool" with a version constraint of 5.12.0, used for semantic versioning based on git history.